+++ R/AFM.getSampleImages.R
@@ -2,7 +2,8 @@
 #'
 #' @description
 #' returns sample AFM images included in this library
-#' @param type could be "ibw" for Asylum Research, "nid" for NanoSurf, "tiff" for Park, if empty or "*", all files will be returned
+#' @param type could be "ibw" for Asylum Research, "nid" for NanoSurf, "tiff" for Park,
+#' if empty or "*", all files will be returned
 #' @return vector with path/filename to AFM sample images
 #' @author Thomas Gredig
 #' @examples
@@ -11,6 +12,7 @@
 #' @export
 AFM.getSampleImages <- function(type='*') {
   pfad = system.file("extdata",package="nanoscopeAFM")
-  file.list = dir(system.file("extdata",package="nanoscopeAFM"), pattern=paste0(type,"$"))
+  file.list = dir(system.file("extdata",package="nanoscopeAFM"),
+                  pattern=paste0(type,"$"))
   file.path(pfad, file.list)
 }