+++ tests/testthat/test-general.R
@@ -1,10 +1,15 @@
 context("general AFM test")
 
+filename = AFM.getSampleImages(type='ibw')
+afmd = AFM.import(filename)
+
+
 
 test_that("AFMdata object", {
   a = AFMdata(instrument='Cypher')
   expect_true(isS4(a))
 })
+
 
 test_that("AFM.flatten: flattening image", {
   filename = system.file("extdata", "NanoSurf_20160301.nid",package="nanoscopeAFM")
@@ -20,19 +25,39 @@ test_that("AFM.getSampleImages: find sample files", {
 })
 
 test_that("line Profile", {
-  filename = AFM.getSampleImages(type='ibw')
-  d = AFM.import(filename)
-  AFM.lineProfile(d, 0,0, 2000,2000) -> d1
+  AFM.lineProfile(afmd, 0,0, 2000,2000) -> d1
   AFM.lineProfile(d1, 0,0, 100,2500) -> d2
   q = AFM.linePlot(d2,dataOnly=TRUE)
   expect_equal(sum(q$z), 128.764, tolerance = 1e-4)
   expect_equal(nlevels(as.factor(q$type)), 2)
 })
 
+test_that("print AFMdata", {
+  expect_output(print(afmd), "Cypher AFM Image")
+  expect_output(print(afmd), "HeightRetrace")
+})
 
 
-test_that("test histogram for image", {
-  filename = AFM.getSampleImages(type='ibw')
-  d = AFM.import(filename)
-  expect_equal( sum(AFM.histogram(d, dataOnly=TRUE)$zDensity), 1)
+test_that("summary AFMinfo", {
+  h = AFMinfo(AFM.getSampleImages(type='ibw')[1])
+  expect_output(summary(h), "Scan Angle")
+})
+
+
+
+test_that("Graphing ggplot2 AFM image", {
+  p = plot(afmd)
+  expect_equal(class(p$layers[[1]]$geom), c("GeomRaster", "Geom","ggproto","gg"))
+  p = plot(afmd, graphType = 2)
+  expect_equal(class(p$layers[[1]]$geom), c("GeomRaster", "Geom","ggproto","gg"))
+  p = plot(afmd, graphType = 3)
+  expect_equal(class(p$layers[[1]]$geom), c("GeomRaster", "Geom","ggproto","gg"))
+})
+
+
+
+test_that("test histogram for image and data", {
+  expect_equal( sum(AFM.histogram(afmd, dataOnly=TRUE)$zDensity), 1)
+  p = AFM.histogram(afmd)
+  expect_equal(class(p$layers[[1]]$geom), c("GeomBar", "GeomRect", "Geom","ggproto","gg"))
 })