+++ R/AFMinfo.R
@@ -19,7 +19,10 @@ AFMinfo <- function(filename) {
     heightPixel = data$value[grep('ScanLines',data$name)[1]]
     scanAngle = data$value[grep('ScanAngle',data$name)[1]]
     scanRate.Hz = data$value[grep('ScanRate',data$name)[1]]
-    note = data$value[grep('ImageNote',data$name)[1]]
+    imagingMode = data$value[grep('ImagingMode',data$name)[1]]
+    cantilever = data$value[grep('ThermalLever',data$name)[1]]
+    resFrequency = data$value[grep('ResFreq1',data$name)[1]]
+    note = paste0(data$value[grep('ImageNote',data$name)[1]],",",data$value[grep('BaseName',data$name)[1]])
     noChannels = 8 - sum(data$value[grep('^Channel\\d',data$name)]==" None")
   } else if (fext=='tiff') {
     data = read.Park_header.v2(filename)
@@ -28,7 +31,10 @@ AFMinfo <- function(filename) {
     heightPixel = data$value[grep('nHeight',data$name)[1]]
     scanAngle = data$value[grep('dfAngle',data$name)[1]]
     scanRate.Hz = data$value[grep('dfScanRateHz',data$name)[1]]
-    note = data$value[grep('ImageNote',data$name)[1]]
+    imagingMode = data$value[grep('imageMode',data$name)[1]]
+    cantilever = data$value[grep('Cantilever',data$name)[1]]
+    resFrequency = data$value[grep('dfNCMFrequency',data$name)[1]]
+    note = data$value[grep('sourceName',data$name)[1]]
     noChannels = 1
   } else if (fext=='nid') {
     data = read.NanoSurf_header.v2(filename)
@@ -37,27 +43,36 @@ AFMinfo <- function(filename) {
     heightPixel = data$value[grep('Lines',data$name)[1]]
     scanAngle = data$value[grep('dfAngle',data$name)[1]]
     scanRate.Hz = 1/as.numeric(gsub('(\\d+).*','\\1',data$value[grep('Time/Line',data$name)[1]]))
-    vibrationFreq.Hz = as.numeric(gsub('kHz','',data$value[grep('Vibration freq',data$name)[1]]))*1000
+    imagingMode = data$value[grep('Op. mode',data$name)[1]]
+    cantilever = data$value[grep('Cantilever type',data$name)[1]]
+    resFrequency = as.numeric(gsub('kHz','',data$value[grep('Vibration freq',data$name)[1]]))*1000
     note = ""
-    noChannels = NA
-  } else {  # Veeco
+    noChannels = length(grep('Gr\\d-Ch\\d',data$name))
+  } else {
     data = read.Nanoscope_file(filename, headerOnly=TRUE)
     type = 'Veeco'
     widthPixel = as.numeric(data$value[grep('Valid data len X',data$name)[1]])
     heightPixel = as.numeric(data$value[grep('Valid data len Y',data$name)[1]])
     scanAngle = data$value[grep('Feature scan angle',data$name)[1]]
     scanRate.Hz = data$value[grep('Scan rate',data$name)[1]]
+    imagingMode = ''
+    cantilever = data$value[grep('Cantilever type',data$name)[1]]
+    resFrequency = 0
     note = data$value[grep('Note',data$name)[1]]
-    noChannels = NA
+    noChannels = 0
   }
   structure(
     list(
       data = data,
       type = type,
       filename = filename,
-      widthPixel = widthPixel,
-      heightPixel = heightPixel,
-      scanRate.Hz = scanRate.Hz,
+      widthPixel = as.numeric(widthPixel),
+      heightPixel = as.numeric(heightPixel),
+      scanRate.Hz = as.numeric(scanRate.Hz),
+      imagingMode = imagingMode,
+      cantilever = cantilever,
+      resFrequency = as.numeric(resFrequency),
+      noChannels = noChannels,
       note = note
     ),
     class = 'AFMinfo'